Move SUFRAMAValidator class to Validation path
