+++ DocsBr/DocsBr.csproj
@@ -82,6 +82,7 @@
     <Compile Include="Validation\IE\IETocantinsValidator.cs" />
     <Compile Include="Validation\IEValidator.cs" />
     <Compile Include="Validation\IE\IIEValidator.cs" />
+    <Compile Include="Validation\SUFRAMAValidator.cs" />
   </ItemGroup>
   <Import Project="$(MSBuildToolsPath)\Microsoft.CSharp.targets" />
   <!-- To modify your build process, add your task inside one of the targets below and uncomment it.
